feat: add validation dialog when delete item

+++ apps/my-app/src/app/pages/user/services/user-list.store.service.ts
@@ -1,10 +1,14 @@
 import MINIMAL_USER_RESPONSE_TABLE_FILTER_OPTION from '@my-app/core/constants/table-filter-option/minimal-user-response.table-filter-option';
+import { AppValidationDialog } from '@my-app/shared/dialog/validation.dialog';
 
 import { AdditionalFlagsRequest } from '@/core/api/additional-flags/additional-flags.request';
 import { AuthService } from '@/core/api/auth/auth.service';
 import { MinimalUserResponse } from '@/core/api/user/user.response';
 import { UserService } from '@/core/api/user/user.service';
+import { APP_TRANSLATION_SERVICE } from '@/core/translation.service';
 import { paginationContainer } from '@/shared/pagination-container';
+
+import { NgpDialogManager } from 'ng-primitives/dialog';
 
 import { inject, Injectable, signal } from '@angular/core';
 
@@ -12,6 +16,8 @@ import { inject, Injectable, signal } from '@angular/core';
 export class UserListStoreService {
   private readonly _authService = inject(AuthService);
   private readonly _userService = inject(UserService);
+  private readonly _dialogManager = inject(NgpDialogManager);
+  private readonly _translationService = inject(APP_TRANSLATION_SERVICE);
 
   public readonly userFilterOption = MINIMAL_USER_RESPONSE_TABLE_FILTER_OPTION;
 
@@ -26,8 +32,18 @@ export class UserListStoreService {
   });
 
   public deleteUser(id: string): void {
-    this._userService.deleteUserById(id).subscribe({
-      next: () => this.usersResource.resource.reload(),
+    AppValidationDialog.open(this._dialogManager, {
+      title: this._translationService.instant('validation.user_delete.title'),
+      description: this._translationService.instant('validation.user_delete.description', {
+        id: id,
+      }),
+      type: 'ok_cancel',
+    }).subscribe((result) => {
+      if (result === 'ok') {
+        this._userService.deleteUserById(id).subscribe({
+          next: () => this.usersResource.resource.reload(),
+        });
+      }
     });
   }
 }
